backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
openai
pydantic-settings
uuid
langchain
langchain-openai
langchain-community
beautifulsoup4
requests
httpx
qdrant-client
sqlalchemy
psycopg[binary]
aiomysql
redis
alembic
cryptography
pypdf
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-logging
opentelemetry-exporter-otlp
python-json-logger
stripe
pytest
